backend/requirements.txt
django
djangorestframework
djoser
djangorestframework_simplejwt
django-extensions
django-cors-headers
python-dotenv
psycopg2